Extract benchmark regression detector into core microcrate
