+++ Cargo.toml
@@ -103,6 +103,7 @@ members = [
   "crates/bitnet-wgpu",          # WebGPU compute backend for cross-platform GPU inference
   "crates/bitnet-dispatch-core", # SRP: backend-agnostic compute dispatch sizing/contracts
   "crates/bitnet-bench-receipts", # SRP: benchmark receipt model + JSONL persistence
+  "crates/bitnet-bench-regression-core", # SRP: reusable benchmark regression detection core
   "crates/bitnet-wgpu-bench",    # Benchmarking and profiling harness for wgpu compute kernels
   "crates/bitnet-wgpu-runner",   # WebGPU kernel runner for GPU compute dispatch
   "crossval",
