+++ crates/bitnet-bench-regression-core/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "bitnet-bench-regression-core"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "Reusable benchmark regression detection primitives"
+rust-version.workspace = true
+
+[dependencies]
+bitnet-bench-receipts = { path = "../bitnet-bench-receipts" }
+
+[lints]
+workspace = true
