+++ crates/bitnet-wgpu-bench/Cargo.toml
@@ -8,10 +8,8 @@ license = "MIT OR Apache-2.0"
 
 [dependencies]
 bitnet-bench-receipts = { path = "../bitnet-bench-receipts" }
+bitnet-bench-regression-core = { path = "../bitnet-bench-regression-core" }
 bitnet-nvidia = { path = "../bitnet-nvidia" }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-thiserror = "2"
 
 [dev-dependencies]
 proptest = "1"
